* tests separated to shared/js/jvm - methodName fixes for scala.js
