+++ core_tests/jvm/src/test/scala/com.paulbutcher.test/proxy/ProxyMockTest.scala
