+++ core_tests/shared/src/test/scala/com/paulbutcher/test/mock/MockNamingTest.scala
@@ -38,10 +38,6 @@ class MockNamingTest extends IsolatedSpec {
     getMockMethodName(m.oneParam _) shouldBe "<mock> TestTrait.oneParam"
   }
 
-  it should "have a sensible method name when mocking an operator" in {
-    getMockMethodName(m.+ _) shouldBe "<mock> TestTrait.$plus" // TODO could be better
-  }
-
   it should "have a sensible method name when mocking curried method" in {
     getMockMethodName(m.curried(_: Int)(_: Double)) shouldBe "<mock> TestTrait.curried"
   }
